PC-1344: extract eligible postcode list to a new service
